requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
langchain-ollama>=0.3.0
streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.20.0
openpyxl>=3.1.0
fpdf2>=2.7.0
tabulate>=0.9.0
pytest>=8.0.0
